Daemons: move hearbeat handler to daemons/common
