+++ lib/rucio/daemons/conveyor/common.py
@@ -21,24 +21,21 @@ from __future__ import division
 
 import datetime
 import logging
-import os
-import socket
-import threading
 import time
 from typing import TYPE_CHECKING
 
 from rucio.common.config import config_get, config_get_int
 from rucio.common.exception import (InvalidRSEExpression, TransferToolTimeout, TransferToolWrongAnswer, RequestNotFound,
                                     DuplicateFileTransferSubmission, VONotFound)
-from rucio.common.logging import formatted_logger
 from rucio.common.utils import PriorityQueue
-from rucio.core import heartbeat, request as request_core, transfer as transfer_core
+from rucio.core import request as request_core, transfer as transfer_core
 from rucio.core.monitor import record_counter, record_timer
 from rucio.core.replica import add_replicas, tombstone_from_delay, update_replica_state
 from rucio.core.request import set_request_state, queue_requests
 from rucio.core.rse import list_rses
 from rucio.core.rse_expression_parser import parse_expression
 from rucio.core.vo import list_vos
+from rucio.daemons.common import HeartbeatHandler
 from rucio.db.sqla import models
 from rucio.db.sqla.constants import RequestState, RequestType, ReplicaState
 from rucio.db.sqla.session import transactional_session
@@ -49,64 +46,6 @@ if TYPE_CHECKING:
     from rucio.core.transfer import DirectTransferDefinition
     from rucio.transfertool.transfertool import Transfertool, TransferToolBuilder
     from sqlalchemy.orm import Session
-
-
-class HeartbeatHandler:
-    """
-    Simple contextmanager which sets a heartbeat and associated logger on entry and cleans up the heartbeat on exit.
-    """
-
-    def __init__(self, executable, renewal_interval, logger_prefix=None):
-        """
-        :param executable: the executable name which will be set in heartbeats
-        :param renewal_interval: the interval at which the heartbeat will be renewed in the database.
-        Calls to live() in-between intervals will re-use the locally cached heartbeat.
-        :param logger_prefix: the prefix to be prepended to all log messages
-        """
-        self.executable = executable
-        self.renewal_interval = renewal_interval
-        self.older_than = renewal_interval * 10 if renewal_interval and renewal_interval > 0 else None  # 10 was chosen without any particular reason
-        self.logger_prefix = logger_prefix or executable
-
-        self.hostname = socket.getfqdn()
-        self.pid = os.getpid()
-        self.hb_thread = threading.current_thread()
-
-        self.logger = None
-        self.last_heart_beat = None
-        self.last_time = None
-
-    def __enter__(self):
-        heartbeat.sanity_check(executable=self.executable, hostname=self.hostname)
-        self.live()
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        if self.last_heart_beat:
-            heartbeat.die(self.executable, self.hostname, self.pid, self.hb_thread)
-            if self.logger:
-                self.logger(logging.INFO, 'Heartbeat cleaned up')
-
-    def live(self):
-        """
-        :return: a tuple: <the number of the current worker>, <total number of workers>, <decorated logger>
-        """
-        if not self.last_time or self.last_time < datetime.datetime.now() - datetime.timedelta(seconds=self.renewal_interval):
-            if self.older_than:
-                self.last_heart_beat = heartbeat.live(self.executable, self.hostname, self.pid, self.hb_thread, older_than=self.older_than)
-            else:
-                self.last_heart_beat = heartbeat.live(self.executable, self.hostname, self.pid, self.hb_thread)
-
-            prefix = '%s[%i/%i]: ' % (self.logger_prefix, self.last_heart_beat['assign_thread'], self.last_heart_beat['nr_threads'])
-            self.logger = formatted_logger(logging.log, prefix + '%s')
-
-            if not self.last_time:
-                self.logger(logging.DEBUG, 'First heartbeat set')
-            else:
-                self.logger(logging.DEBUG, 'Heartbeat renewed')
-            self.last_time = datetime.datetime.now()
-
-        return self.last_heart_beat['assign_thread'], self.last_heart_beat['nr_threads'], self.logger
 
 
 def run_conveyor_daemon(once, graceful_stop, executable, logger_prefix, partition_wait_time, sleep_time, run_once_fnc, activities=None):
